feat: validate inventory storage boundaries
- Validate inventory JSON with the Inventory RootModel on load.
- Validate inventory state before saving updates.
- Keep inventory service callers working with plain dictionaries.
- Reduce repeated inventory saves during reservation and sale finalization.

+++ app/models/inventory.py
@@ -7,7 +7,7 @@ class InventoryItem(BaseModel):
     )
 
     name: str
-    price: int = Field(ge=0)
+    price: float = Field(ge=0)
     available_stock: int = Field(ge=0)
     reserved_stock: int = Field(ge=0)
     sold_stock: int = Field(ge=0)
